# Optimize assembly/namespace pair creation and updating in XamlSchema