--- a/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/MS/Impl/AssemblyNamespacePair.cs
+++ b/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/MS/Impl/AssemblyNamespacePair.cs
@@ -1,17 +1,16 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-#nullable disable
-
+using System.Diagnostics;
 using System.Reflection;
 
 namespace System.Xaml.MS.Impl
 {
     [DebuggerDisplay("{ClrNamespace} {Assembly.FullName}")]
-    internal class AssemblyNamespacePair
+    internal readonly struct AssemblyNamespacePair
     {
-        private WeakReference _assembly;
-        private string _clrNamespace;
+        private readonly WeakReference _assembly;
+        private readonly string _clrNamespace;
 
         public AssemblyNamespacePair(Assembly asm, string clrNamespace)
         {
@@ -19,9 +18,9 @@ namespace System.Xaml.MS.Impl
             _clrNamespace = clrNamespace;
         }
 
-        public Assembly Assembly
+        public Assembly? Assembly
         {
-            get { return (Assembly)_assembly.Target; }
+            get { return (Assembly?)_assembly.Target; }
         }
 
         public string ClrNamespace