--- a/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/Schema/XamlNamespace.cs
+++ b/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/Schema/XamlNamespace.cs
@@ -237,13 +237,9 @@ namespace System.Xaml.Schema
         {
             Assembly asm = SchemaContext.OnAssemblyResolve(assemblyName);
             if (asm is null)
-            {
                 return null;
-            }
 
-            List<AssemblyNamespacePair> onePair = new List<AssemblyNamespacePair>();
-            onePair.Add(new AssemblyNamespacePair(asm, clrNs));
-            return onePair;
+            return new List<AssemblyNamespacePair>(1) { new AssemblyNamespacePair(asm, clrNs) };
         }
 
         private Type SearchAssembliesForShortName(string shortName)