--- a/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/XamlSchemaContext.cs
+++ b/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/XamlSchemaContext.cs
@@ -1182,15 +1182,15 @@ namespace System.Xaml
 
         private bool UpdateNamespaceByUriList(XmlNsInfo nsInfo)
         {
-            bool foundNew = false;
             IList<XmlNsInfo.XmlNsDefinition> xmlnsDefs = nsInfo.NsDefs;
-            int xmlnsDefsCount = xmlnsDefs.Count;
-            for (int i = 0; i < xmlnsDefsCount; i++)
+            bool foundNew = false;
+
+            for (int i = 0; i < xmlnsDefs.Count; i++)
             {
                 XmlNsInfo.XmlNsDefinition xmlnsDef = xmlnsDefs[i];
-                AssemblyNamespacePair pair = new AssemblyNamespacePair(nsInfo.Assembly, xmlnsDef.ClrNamespace);
                 XamlNamespace ns = GetXamlNamespace(xmlnsDef.XmlNamespace);
-                ns.AddAssemblyNamespacePair(pair);
+
+                ns.AddAssemblyNamespacePair(new AssemblyNamespacePair(nsInfo.Assembly, xmlnsDef.ClrNamespace));
                 foundNew = true;
             }