--- a/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/XamlSchemaContext.cs
+++ b/src/Microsoft.DotNet.Wpf/src/System.Xaml/System/Xaml/XamlSchemaContext.cs
@@ -1180,6 +1180,7 @@ namespace System.Xaml
             return foundNew;
         }
 
+        // This method should be called inside _syncExaminingAssemblies lock
         private bool UpdateNamespaceByUriList(XmlNsInfo nsInfo)
         {
             IList<XmlNsInfo.XmlNsDefinition> xmlnsDefs = nsInfo.NsDefs;