## Description

Adjusts inefficient `List<AssemblyNamespacePair>` copies in `AddAssemblyNamespacePair` that result in up to double gen0 allocations needed for the job, offering perf benefit and decreased allocations.

- I have changed it to be a `readonly struct`, there's like zero benefit but I think it serves better purpose.
- The `WeakReference` is now generic to prevent unnecessary casting.
- Since `_assemblyNamespaces` is under a write-lock and copies are created each time to allow reading concurrently by multiple readers, we can just swap to an array for overall efficiency since only 1 item is added each time.

### Initial creation
| Method   | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | 108.37 ns |   2.179 ns |    2.594 ns | 0.0086 |         317 B |         144 B |
| PR_EDIT |  89.78 ns |   1.833 ns |    2.687 ns | 0.0038 |         223 B |          64 B |

### Initial creation plus 9x AddAssemblyNamespacePair

| Method   | Mean [ns]  | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |-----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original |  1,419.8 ns |   28.21 ns |    37.66 ns | 0.1678 |       2,668 B |        2824 B |
| PR_EDIT |  1,100.8 ns |   18.28 ns |    16.20 ns | 0.0935 |       2,329 B |        1584 B |

## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build, basic method testing.

## Risk

Low, there's only a swap from class to readonly struct being stored in a private array (that was originally a list).

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9926)