@dipeshmsft 2025-05-16
@h3xds1nz, this PR looks good to me. Can you resolve the merge conflicts here ?

**PS:** From what I gather, copying a List is heavier as compared to array and that's why the shift ? If yes, then maybe we can add this as a comment where AssemblyNamespacePair[] array is defined.