@dipeshmsft 2025-05-16
> This means you're here initializing the List<T> from another List<T>, which creates a precisely-sized array onto you which then add a new item which incurs a new, practically redundant, copy of the just freshly created List<T>.

Okay, so now using array, we pre-allocate the extra space for the 1 extra element, preventing the extra copy. Thanks a lot for the explaination.